Fix MongoDB connection caching
