+++ database/dbConnect.ts
@@ -1,0 +1,47 @@
+import mongoose from "mongoose"
+
+const MONGODB_URI = process.env.MONGODB_URI
+
+if (!MONGODB_URI) {
+  throw new Error(
+    "Please define the MONGODB_URI environment variable inside .env.local"
+  )
+}
+
+type MongooseConnection = {
+  conn: typeof mongoose | null
+  promise: Promise<typeof mongoose> | null
+}
+
+declare global {
+  // eslint-disable-next-line no-var
+  var mongooseConnection: MongooseConnection | undefined
+}
+
+const cached: MongooseConnection =
+  global.mongooseConnection ?? (global.mongooseConnection = { conn: null, promise: null })
+
+async function dbConnect() {
+  if (cached.conn) {
+    return cached.conn
+  }
+
+  if (!cached.promise) {
+    const opts = {
+      bufferCommands: false,
+    }
+
+    cached.promise = mongoose.connect(MONGODB_URI, opts)
+  }
+
+  try {
+    cached.conn = await cached.promise
+  } catch (error) {
+    cached.promise = null
+    throw error
+  }
+
+  return cached.conn
+}
+
+export default dbConnect
